WIIS-6220: Fix preparation

+++ src/Service/PreparationsManagerService.php
@@ -196,6 +196,10 @@ class PreparationsManagerService
             $line = $preparationOrderArticleLineRepository->find($lineId);
             $newPreparation->addArticleLine($line);
             $preparation->removeArticleLine($line);
+            if($line->getPreparation() === null) {
+                $entityManager->remove($line);
+            }
+
             $articleToKeep = $line->getArticle();
             $articleToKeep
                 ->setStatut($statutRepository->findOneByCategorieNameAndStatutCode(CategorieStatut::ARTICLE, Article::STATUT_EN_TRANSIT));
@@ -444,7 +448,10 @@ class PreparationsManagerService
                                 $splitArticleLineIds[] = $line->getId();
                             } else {
                                 $preparation->removeArticleLine($line);
-                                $entityManager->remove($line);
+                                if($line->getPreparation() === null) {
+                                    $entityManager->remove($line);
+                                }
+
                                 $article->setStatut($articleActiveStatus);
                             }
                             $article->setQuantite($article->getQuantite() - $pickedQuantity);
@@ -480,6 +487,10 @@ class PreparationsManagerService
                         }
                         else {
                             $preparation->removeArticleLine($line);
+                            if($line->getPreparation() === null) {
+                                $entityManager->remove($line);
+                            }
+
                             $splitArticleLineIds[] = $line->getId();
                         }
                     }
